requirements.txt
disnake == 2.5.1
async_google_trans_new
python-aiml
python-dotenv